+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 ### Added
 
+### Fixed 
+- `Line.IsCollinear(Line line)` would return `false` if lines are close to each other but not collinear 
+- `Vector3.AreCollinear(Vector3 a, Vector3 b, Vector3 c)` would return `false` if points coordinates difference is larger than `Vector3.EPSILON`
 - `EdgeDisplaySettings` for materials to control the display of lines in supported viewers (like Hypar.io).
 - `Line.MergeCollinearLine(Line line)` creates new line containing all four collinear vertices 
 
 ### Changed
-
 - Simplified `IEnumerable<Vector3>.ToGraphicsBuffers()`
 - `TryToGraphicsBuffers` is now public